split up stats_impl_test to match the *impl.h and and *impl.cc files.
