+++ test/common/stats/BUILD
@@ -9,16 +9,54 @@ load(
 envoy_package()
 
 envoy_cc_test(
-    name = "stats_impl_test",
-    srcs = ["stats_impl_test.cc"],
+    name = "heap_stat_data_test",
+    srcs = ["heap_stat_data_test.cc"],
     deps = [
-        "//source/common/common:hex_lib",
         "//source/common/stats:heap_stat_data_lib",
-        "//source/common/stats:stats_lib",
-        "//test/mocks/stats:stats_mocks",
+        "//source/common/stats:stats_options_lib",
+        # "//test/mocks/stats:stats_mocks",
         "//test/test_common:logging_lib",
+        #        "//test/test_common:utility_lib",
+        #        "@envoy_api//envoy/config/metrics/v2:stats_cc",
+    ],
+)
+
+envoy_cc_test(
+    name = "isolated_store_impl_test",
+    srcs = ["isolated_store_impl_test.cc"],
+    deps = [
+        "//source/common/stats:isolated_store_lib",
+        #"//test/mocks/stats:stats_mocks",
+    ],
+)
+
+envoy_cc_test(
+    name = "source_impl_test",
+    srcs = ["source_impl_test.cc"],
+    deps = [
+        "//source/common/stats:source_impl_lib",
+        "//test/mocks/stats:stats_mocks",
+    ],
+)
+
+envoy_cc_test(
+    name = "tag_extractor_test",
+    srcs = ["tag_extractor_test.cc"],
+    deps = [
+        "//source/common/stats:tag_extractor_lib",
+        "//source/common/stats:tag_producer_lib",
+        #        "//test/mocks/stats:stats_mocks",
         "//test/test_common:utility_lib",
-        "@envoy_api//envoy/config/metrics/v2:stats_cc",
+    ],
+)
+
+envoy_cc_test(
+    name = "tag_producer_test",
+    srcs = ["tag_producer_test.cc"],
+    deps = [
+        "//source/common/stats:tag_producer_lib",
+        #        "//test/mocks/stats:stats_mocks",
+        "//test/test_common:utility_lib",
     ],
 )
