+++ test/common/stats/isolated_store_impl_test.cc
@@ -1,0 +1,100 @@
+#include <string>
+
+#include "envoy/stats/stats_macros.h"
+
+#include "common/stats/isolated_store_impl.h"
+
+#include "absl/strings/str_cat.h"
+#include "absl/strings/string_view.h"
+#include "gtest/gtest.h"
+
+namespace Envoy {
+namespace Stats {
+
+TEST(StatsIsolatedStoreImplTest, All) {
+  IsolatedStoreImpl store;
+
+  ScopePtr scope1 = store.createScope("scope1.");
+  Counter& c1 = store.counter("c1");
+  Counter& c2 = scope1->counter("c2");
+  EXPECT_EQ("c1", c1.name());
+  EXPECT_EQ("scope1.c2", c2.name());
+  EXPECT_EQ("c1", c1.tagExtractedName());
+  EXPECT_EQ("scope1.c2", c2.tagExtractedName());
+  EXPECT_EQ(0, c1.tags().size());
+  EXPECT_EQ(0, c1.tags().size());
+
+  Gauge& g1 = store.gauge("g1");
+  Gauge& g2 = scope1->gauge("g2");
+  EXPECT_EQ("g1", g1.name());
+  EXPECT_EQ("scope1.g2", g2.name());
+  EXPECT_EQ("g1", g1.tagExtractedName());
+  EXPECT_EQ("scope1.g2", g2.tagExtractedName());
+  EXPECT_EQ(0, g1.tags().size());
+  EXPECT_EQ(0, g1.tags().size());
+
+  Histogram& h1 = store.histogram("h1");
+  Histogram& h2 = scope1->histogram("h2");
+  scope1->deliverHistogramToSinks(h2, 0);
+  EXPECT_EQ("h1", h1.name());
+  EXPECT_EQ("scope1.h2", h2.name());
+  EXPECT_EQ("h1", h1.tagExtractedName());
+  EXPECT_EQ("scope1.h2", h2.tagExtractedName());
+  EXPECT_EQ(0, h1.tags().size());
+  EXPECT_EQ(0, h2.tags().size());
+  h1.recordValue(200);
+  h2.recordValue(200);
+
+  ScopePtr scope2 = scope1->createScope("foo.");
+  EXPECT_EQ("scope1.foo.bar", scope2->counter("bar").name());
+
+  // Validate that we sanitize away bad characters in the stats prefix.
+  ScopePtr scope3 = scope1->createScope(std::string("foo:\0:.", 7));
+  EXPECT_EQ("scope1.foo___.bar", scope3->counter("bar").name());
+
+  EXPECT_EQ(4UL, store.counters().size());
+  EXPECT_EQ(2UL, store.gauges().size());
+}
+
+TEST(StatsIsolatedStoreImplTest, LongStatName) {
+  IsolatedStoreImpl store;
+  Stats::StatsOptionsImpl stats_options;
+  const std::string long_string(stats_options.maxNameLength() + 1, 'A');
+
+  ScopePtr scope = store.createScope("scope.");
+  Counter& counter = scope->counter(long_string);
+  EXPECT_EQ(absl::StrCat("scope.", long_string), counter.name());
+}
+
+/**
+ * Test stats macros. @see stats_macros.h
+ */
+// clang-format off
+#define ALL_TEST_STATS(COUNTER, GAUGE, HISTOGRAM)                                                  \
+  COUNTER  (test_counter)                                                                          \
+  GAUGE    (test_gauge)                                                                            \
+  HISTOGRAM(test_histogram)
+// clang-format on
+
+struct TestStats {
+  ALL_TEST_STATS(GENERATE_COUNTER_STRUCT, GENERATE_GAUGE_STRUCT, GENERATE_HISTOGRAM_STRUCT)
+};
+
+TEST(StatsMacros, All) {
+  IsolatedStoreImpl stats_store;
+  TestStats test_stats{ALL_TEST_STATS(POOL_COUNTER_PREFIX(stats_store, "test."),
+                                      POOL_GAUGE_PREFIX(stats_store, "test."),
+                                      POOL_HISTOGRAM_PREFIX(stats_store, "test."))};
+
+  Counter& counter = test_stats.test_counter_;
+  EXPECT_EQ("test.test_counter", counter.name());
+
+  Gauge& gauge = test_stats.test_gauge_;
+  EXPECT_EQ("test.test_gauge", gauge.name());
+
+  Histogram& histogram = test_stats.test_histogram_;
+  EXPECT_EQ("test.test_histogram", histogram.name());
+}
+
+} // namespace Stats
+} // namespace Envoy
